feat: Refactor volunteer-related functions into volunteer-handlers.ts and update imports
